fix(directive): change closable directive for an OutsideClick component
