+++ src/components/Editor.vue
@@ -12,8 +12,6 @@
     <StageBar v-on:change-stage="changeStage" />
     <RadonStage class="stage" :stage="currentStage" :script="currentScript" />
     <Console />
-    <p ref="delete-ref"></p>
-    <p v-closable="{ exclude: ['delete-ref'], handler: 'onClose' }"></p>
   </div>
 </template>
 
@@ -37,8 +35,6 @@ export default {
   },
   data() {
     return {
-      // TODO: remove ref delete-ref and directive associated. Fix directive to find last reference
-      close: false,
       currentStage: 'retrieve',
     }
   },
@@ -48,9 +44,6 @@ export default {
     },
     clearError(errorName) {
       this.$store.commit('clearError', { error: errorName })
-    },
-    onClose() {
-      this.close = false
     },
   },
   computed: {
