+++ src/components/RadonOperator.vue
@@ -1,6 +1,6 @@
 <template>
   <div class="radon-operator">
-    <FormSelect
+    <Select
       :value="selectedOption"
       @input="option => updateTemplate(selectedOption.id, option.value)"
       :options="operatorOptions"
@@ -22,7 +22,7 @@
           v-if="argument.markup_type === 'select'"
         >
           <p>{{ argument.label }}</p>
-          <FormSelect
+          <Select
             class="argument-options"
             :value="selectedArgument[index]"
             :options="argumentOptions[index]"
@@ -41,8 +41,7 @@
 </template>
 
 <script>
-import Select from '@/components/Select'
-import FormSelect from './FormSelect'
+import Select from './Select'
 import Input from '@/components/Input'
 import { UPDATE_TEMPLATE, STORE_VARIABLE_COMPONENT_ID } from '@/store/mutation-types'
 
@@ -61,7 +60,6 @@ export default {
   components: {
     Input,
     Select,
-    FormSelect,
   },
   methods: {
     updateTemplate(id, value, variables) {
@@ -87,7 +85,6 @@ export default {
   },
   computed: {
     selectedOption() {
-      console.log('>>>>', this.operator.selected)
       return {
         id: this.operator.id,
         primaryText: this.operator.selected.label,
